feat(): update parallel connections and duration settings

+++ download.go
@@ -42,7 +42,7 @@ func startDownloadTest(targets []Target) float64 {
 		totalTime += result.TotalTime
 	}
 
-	downloadSpeed := (float64(totalSize * uint64(time.Second))) / float64(totalTime) * parallelConnections
+	downloadSpeed := (float64(totalSize * uint64(time.Second))) / float64(totalTime) * float64(parallelConnections)
 
 	log.Debug().Dur("totalTime", totalTime).Uint64("totalSize", totalSize).Float64("speed", downloadSpeed/1e3/1e3*8).Msg("DownloadSpeed")
